Only set new jobs to accepted
