+++ medicines/doc-index-updater/src/state_manager/script.lua
@@ -1,0 +1,5 @@
+if redis.call("EXISTS", KEYS[1]) == 0 then
+    return redis.call("SET", KEYS[1], "Accepted")
+else
+    return nil
+end
